FisherTest_GO_BP_MF_CC {seq2pathway}	R Documentation

A wrapper function to perform the Fisher's exact test, using GO-defined genesets.

Description

A wrapper function to perform the Fisher's exact test, using GO-defined genesets.

Usage


FisherTest_GO_BP_MF_CC(gs, genome=c("hg38","hg19","mm10","mm9"), 
                       min_Intersect_Count=5,
                       Ontology=c("GOterm", "BP","MF", "CC","newOntology"), 
			  newOntology=NULL)

Arguments

gs
A characteristic vector of gene symbols, the input gene list.

genome
A character specifies the genome type. Currently, choice of "hg38", "hg19", "mm10", and "mm9" is supported.

min_Intersect_Count
A number decides the cutoff of the minimum number of intersected genes when reporting Fisher's exact tested results.

Ontology
A character specifies the Gene Ontology, choice of "GOterm", "BP","MF", "CC" and "newOntology" is supported.

newOntology
A list of two lists with the same ontology IDs. or each ontology ID, the 1st list is the lists of defined genes and the 2nd list is the desceiption.

Value

A list of 3 data frames, each is a result of Fisher's exact test, using GO CC, BP, MF respectively. Each data frame reports FET results with the following columns.

GOID
GO term IDs

Description
GO definition and description for the gene-sets

Fisher_Pvalue
is the raw P-values

Fisher_odds
estimate of the odds ratios

FDR
the multi-test adjusted P-values using the Benjamini and Hochberg method

Intersect_Count
the sizes of overlap between GO gene members and the input genelist

GO_gene_inBackground
the counts of genes among each GO term that are also within the given genome background

GO_gene_raw_count
the original counts of genes in each GO term

Intersect_gene
the intersecting genes' symbols

Author(s)

Bin Wang, Xinan Yang

Examples


  data(dat_chip)
  head(dat_chip)
  data(GO_BP_list,package="seq2pathway.data")
  data(Des_BP_list,package="seq2pathway.data")
  newOntology <- list(GO_BP_list[1:200], Des_BP_list[1:200])
  # A demo run of this funcion
  FS_test<- FisherTest_GO_BP_MF_CC(gs=as.vector(rownames(dat_chip)),
                                 Ontology="newOntology", newOntology=newOntology)
  FS_test
  
  ## Not run: 
  data(dat_chip)
  FS_test<-FisherTest_GO_BP_MF_CC(gs=rownames(dat_chip)[1:20], genome="hg19", 
           min_Intersect_Count=1, Ontology="BP")
  FS_test$GO_BP[1:3,]
## End(Not run)          

[Package seq2pathway version 1.14.0 Index]