GRanges_demo {seq2pathway}	R Documentation

loci information with GRanges format

Description

loci information with GRanges format

Usage

data("GRanges_demo")

Format

GRanges object with 10 ranges and 3 metadata columns.

Value

GRanges object

References

Lawrence M, Huber W, Pages H, Aboyoun P, Carlson M, Gentleman R, Morgan M and Carey V(2013). "Software for Computing and Annotating Genomic Ranges." PLoS Computational Biology, 9.

Examples


data(GRanges_demo)

[Package seq2pathway version 1.14.0 Index]