addDescription {seq2pathway}	R Documentation

Retrieve "gene description" attributes for gene symbol.

Description

A function wrappered from Rpackage "biomaRt". Get gene description information from gene symbol information.

Usage


addDescription(genome=c("hg19","mm10","mm9"), genevector)

Arguments

genome
A character specifies the genome type. Currently, choice of "hg19", "mm10", and "mm9" is supported.

genevector
A characteristic vector of gene symbols.

Value

A characteristic matrix of gene symbols and descriptions.

Author(s)

Bin Wang

References

Durinck S, Spellman P, Birney E and Huber W (2009) Mapping identifiers for the integration of genomic datasets with the R/Bioconductor package biomaRt. Nature Protocols, 4, 1184–1191.

Durinck S, Moreau Y, Kasprzyk A, Davis S, De Moor B, Brazma A and Huber W (2005) BioMart and Bioconductor: a powerful link between biological databases and microarray data analysis. Bioinformatics, 21, 3439–3440.

Examples


  require(biomaRt)
  data(dat_chip)
  gene_description<-addDescription(genome="hg19",genevector=rownames(dat_chip)[1:3])

[Package seq2pathway version 1.14.0 Index]